finished structs, testing is passing.
